+++ de.ibapl.jnhw.posix/src/main/native/TermiosFunctions.c
@@ -5,6 +5,8 @@
 #include "de_ibapl_jnhw_posix_Termios.h"
 #include <termios.h>
 #include <errno.h>
+#include <stdint.h>
+
 
 #ifdef __cplusplus
 extern "C" {
@@ -17,7 +19,7 @@ extern "C" {
      */
     JNIEXPORT jint JNICALL Java_de_ibapl_jnhw_posix_Termios_cfgetispeed
     (JNIEnv *env, jclass clazz, jlong termiosBaseAddress) {
-        return cfgetispeed((void*)(long)termiosBaseAddress);
+        return cfgetispeed((void*)(uintptr_t)termiosBaseAddress);
     }
 
     /*
@@ -27,7 +29,7 @@ extern "C" {
      */
     JNIEXPORT jint JNICALL Java_de_ibapl_jnhw_posix_Termios_cfgetospeed
     (JNIEnv *env, jclass clazz, jlong termiosBaseAddress) {
-        return cfgetospeed((void*)(long)termiosBaseAddress);
+        return cfgetospeed((void*)(uintptr_t)termiosBaseAddress);
     }
 
     /*
@@ -37,7 +39,7 @@ extern "C" {
      */
     JNIEXPORT jint JNICALL Java_de_ibapl_jnhw_posix_Termios_cfsetispeed
     (JNIEnv *env, jclass clazz, jlong termiosBaseAddress, jint speed) {
-        int result = cfsetispeed((void*)(long)termiosBaseAddress, speed);
+        int result = cfsetispeed((void*)(uintptr_t)termiosBaseAddress, speed);
         if (result < 0) {
             throw_NativeErrorException(env, errno);
         }
@@ -51,7 +53,7 @@ extern "C" {
      */
     JNIEXPORT jint JNICALL Java_de_ibapl_jnhw_posix_Termios_cfsetospeed
     (JNIEnv *env, jclass clazz, jlong termiosBaseAddress, jint speed) {
-        int result = cfsetospeed((void*)(long)termiosBaseAddress, speed);
+        int result = cfsetospeed((void*)(uintptr_t)termiosBaseAddress, speed);
         if (result < 0) {
             throw_NativeErrorException(env, errno);
         }
@@ -93,7 +95,7 @@ extern "C" {
      */
     JNIEXPORT jint JNICALL Java_de_ibapl_jnhw_posix_Termios_tcgetattr
     (JNIEnv *env, jclass clazz, jint fildes, jlong termiosBaseAddress) {
-        int result = tcgetattr(fildes, (void*)(long)termiosBaseAddress);
+        int result = tcgetattr(fildes, (void*)(uintptr_t)termiosBaseAddress);
         if (result < 0) {
             throw_NativeErrorException(env, errno);
         }
@@ -121,7 +123,7 @@ extern "C" {
      */
     JNIEXPORT jint JNICALL Java_de_ibapl_jnhw_posix_Termios_tcsetattr
     (JNIEnv *env, jclass clazz, jint fildes, jint optional_actions, jlong termiosBaseAddress) {
-        int result = tcsetattr(fildes, optional_actions, (void*)(long)termiosBaseAddress);
+        int result = tcsetattr(fildes, optional_actions, (void*)(uintptr_t)termiosBaseAddress);
         if (result < 0) {
             throw_NativeErrorException(env, errno);
         }
@@ -135,7 +137,7 @@ extern "C" {
      */
     JNIEXPORT jint JNICALL Java_de_ibapl_jnhw_posix_Termios_cfsetspeed
     (JNIEnv *env, jclass clazz, jlong termiosBaseAddress, jint speed) {
-        int result = cfsetspeed((void*)(long)termiosBaseAddress, speed);
+        int result = cfsetspeed((void*)(uintptr_t)termiosBaseAddress, speed);
         if (result < 0) {
             throw_NativeErrorException(env, errno);
         }
